+++ src/components/list/existing-list/existing-list.tsx
@@ -1,14 +1,14 @@
 import { BackButton, Button } from "@/components/ui";
-import { Text } from "react-native";
+import { Alert, Text, View } from "react-native";
 import { ListContent } from "../list-content";
-import { type ProductsListData, type ExistingListProps } from "@/constants";
+import { type ProductsListData, type ExistingListProps, ErrorObject } from "@/constants";
 import { createProductsLists, deleteProductListData, fetchProductsLists, updateProductsList } from "@/store/api";
 import { useSelector } from "react-redux";
 import { selectUser } from "@/store/user";
 import { type MainNavigationProp } from "@/navigation";
 import { useNavigation } from "@react-navigation/native";
 import { useAppDispatch } from "@/store";
-import { colorDark } from "@/styles";
+import { colorDark, fontsStyles } from "@/styles";
 import { TrashSvgComponent } from "@/assets";
 import { type Dispatch } from "react";
 
@@ -21,7 +21,7 @@ export const ExistingList = ({
   language,
   isNewList,
   setProductData,
-}: ExistingListProps & { isNewList: boolean; setProductData: Dispatch<ProductsListData> }) => {
+}: ExistingListProps & { isNewList: boolean; setProductData: Dispatch<React.SetStateAction<ProductsListData>> }) => {
   const user = useSelector(selectUser);
   const dispatch = useAppDispatch();
   const id = user?.id;
@@ -29,14 +29,21 @@ export const ExistingList = ({
 
   const hendleSave = async (userId: string | number) => {
     const data = { ...productData, owner_id: userId };
-
     if (isNewList) {
       try {
         await createProductsLists(data);
         dispatch(fetchProductsLists(userId));
         navigation.navigate("Home");
       } catch (error) {
-        console.error("Ошибка при создании нового списка:", error);
+        if (
+          typeof error === "object" &&
+          error !== null &&
+          "statusErr" in error &&
+          (error as ErrorObject).statusErr === 409
+        ) {
+          setProductData({ ...productData, name: "" });
+          Alert.alert("Такой список уже существует");
+        }
       }
     } else {
       if (data.id) {
@@ -52,21 +59,38 @@ export const ExistingList = ({
   };
 
   const hendleDell = async (objId: string | number, userId: string | number) => {
-    try {
-      await deleteProductListData(objId, userId);
-      dispatch(fetchProductsLists(userId));
-      navigation.navigate("Home");
-    } catch (error) {
-      console.error("Ошибка при удалении списка:", error);
-    }
+    Alert.alert(
+      "Confirm Deletion",
+      "Are you sure you want to delete this product?",
+      [
+        {
+          text: "Cancel",
+          onPress: () => {},
+          style: "cancel",
+        },
+        {
+          text: "OK",
+          onPress: async () => {
+            try {
+              await deleteProductListData(objId, userId);
+              dispatch(fetchProductsLists(userId));
+              navigation.navigate("Home");
+            } catch (error) {
+              console.error("Ошибка при удалении списка:", error);
+            }
+          },
+        },
+      ],
+      { cancelable: true },
+    );
   };
 
   return (
     <>
       <BackButton theme={theme} />
-      <Text>List ID: {listId}</Text>
-      <Text>List Name: {listName}</Text>
-      <Button theme={theme} onPress={handleAddClick}>
+      <Text style={(fontsStyles.text, { color: colorDark.textColor })}>List ID: {listId}</Text>
+      <Text style={(fontsStyles.text, { color: colorDark.textColor })}>List Name: {listName}</Text>
+      <Button style={{ marginTop: 25 }} theme={theme} onPress={handleAddClick}>
         <Text>Добавить</Text>
       </Button>
       {productData?.products && (
@@ -86,13 +110,16 @@ export const ExistingList = ({
           {!isNewList && productData && productData?.id ? (
             <Button
               theme={theme}
+              style={{ marginTop: 15, marginBottom: 25 }}
               onPress={() => {
                 productData.id ? hendleDell(productData.id, id) : null;
               }}
             >
               <TrashSvgComponent width={25} height={25} color={colorDark.textColor} />
             </Button>
-          ) : null}
+          ) : (
+            <View style={{ height: 25 }} />
+          )}
         </>
       ) : null}
     </>
